update api deleteimage

+++ src/controllers/product.js
@@ -1,6 +1,12 @@
 const prisma = require("../config/prisma");
 const cloudinary = require("cloudinary").v2;
 
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUNDINARY_API_KEY,
+  api_secret: process.env.CLOUNDINARY_API_SECRET, // Click 'View API Keys' above to copy your API secret
+});
+
 exports.create = async (req, res) => {
   try {
     const { title, description, price, quantity, categoryId, images } =
@@ -104,11 +110,37 @@ exports.update = async (req, res) => {
 exports.remove = async (req, res) => {
   try {
     const { id } = req.params;
+
+    const product = await prisma.product.findFirst({
+      where: {
+        id: Number(id),
+      },
+      include: {
+        images: true,
+      },
+    });
+    if (!product) {
+      return res.status(400).json({ message: "Product not found" });
+    }
+    console.log(product);
+
+    const deleteImageProduct = product.images.map(
+      (image) =>
+        new Promise((resolve, reject) => {
+          cloudinary.uploader.destroy(image.public_id, (error, result) => {
+            if (error) reject(error);
+            else resolve(result);
+          });
+        })
+    );
+    await Promise.all(deleteImageProduct);
+
     await prisma.product.delete({
       where: {
         id: Number(id),
       },
     });
+
     res.send("Deleted Success");
   } catch (error) {
     console.log(error);
@@ -210,12 +242,6 @@ exports.searchFilters = async (req, res) => {
   }
 };
 
-cloudinary.config({
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-  api_key: process.env.CLOUNDINARY_API_KEY,
-  api_secret: process.env.CLOUNDINARY_API_SECRET, // Click 'View API Keys' above to copy your API secret
-});
-
 exports.createImages = async (req, res) => {
   try {
     const result = await cloudinary.uploader.upload(req.body.image, {
@@ -234,7 +260,7 @@ exports.removeImage = async (req, res) => {
     const { public_id } = req.body;
 
     cloudinary.uploader.destroy(public_id, (result) => {
-      res.status(200).json({message : "Remove Image Success"});
+      res.status(200).json({ message: "Remove Image Success" });
     });
   } catch (error) {
     console.log(error);
